コントローラの共通処理を共通モジュールに移行
